http respon

+++ src/main/java/com/example/demo/service/KelompokService.java
@@ -152,8 +152,13 @@ public class KelompokService {
     
     public ResponseEntity<Object> deleteKelompok(Long id) {
             try {
-                log.info("Executing delete kelompok by id: {}", id);
+                log.info("Check by Kelompok id: "+id);
                 Optional<Kelompok> kelompokById = kelompokRepository.findById(id);
+                if(kelompokById.isEmpty()){
+                    log.info("Kelompok id "+id+ "NOT FOUND");
+                    return ResponseUtil.build(AppConstant.ResponseCode.DATA_NOT_FOUND, null, HttpStatus.NOT_FOUND);
+                }
+                log.info("Executing delete kelompok by id: {}", id);
                 kelompokById.ifPresent(res -> {
                     kelompokRepository.delete(res);
                 });
